requirements.txt
streamlit==1.35.0
openai==0.27.6
langchain==0.2.0
PyMuPDF==1.24.2
pdfminer.six==20221105
docx2txt==0.8
python-docx==1.1.0
python-dotenv==1.0.1
pyttsx3==2.90
pillow==10.3.0
chardet==5.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
httpx==0.27.0
